requirements.txt
flask==3.1.2
line-bot-sdk==3.21.0
openai==2.14.0
gunicorn==23.0.0
